sandbox/requirements_enhanced.txt
# Enhanced FinDocPipeline Dependencies
# Streamlit-compatible lightweight dependencies for multi-document processing

# Core Streamlit and data processing
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0

# Current file processing (keep existing)
pdfplumber>=0.9.0
PyMuPDF>=1.23.0
opencv-python-headless>=4.8.0
pytesseract>=0.3.10
Pillow>=10.0.0
regex>=2023.8.8
openpyxl>=3.1.2

# Multi-format file processing
python-docx>=0.8.11
xlrd>=2.0.1
filetype>=1.2.0
chardet>=5.2.0

# Schema validation and data modeling
pydantic>=2.0.0
jsonschema>=4.19.0

# Lightweight ML and NLP (Streamlit-compatible)
scikit-learn>=1.3.0
nltk>=3.8.1
textstat>=0.7.3

# Visualization and analytics
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Performance and monitoring
psutil>=5.9.0
tqdm>=4.65.0

# Utilities
python-dateutil>=2.8.2